requirements.txt
instagrapi==2.1.2
exifread==3.0.0
python-dotenv==1.0.1
Pillow==10.4.0
urllib3==1.26.18
ffmpeg-python==0.2.0